rendered the buttons and made data display

+++ assets/js/script.js
@@ -12,8 +12,10 @@ let mainHum = document.querySelector("#mainHum");
 let weatherSymbol = document.querySelector("#weatherSymbol");
 let cardIcon = document.querySelectorAll(".cardWeather");
 let cardBio = document.querySelectorAll(".font");
-console.log(cardIcon);
-console.log(cardBio);
+let cityInput = document.querySelector("#cityInput");
+
+
+
 
 const cityList = JSON.parse(localStorage.getItem("cities")) || [];
 
@@ -50,6 +52,7 @@ function displaydate() {
 
 function getValue(event) {
   event.preventDefault();
+  event.stopPropagation();
   let searchInputValue = searchInput.value.trim();
 
   if (searchInputValue !== "" && searchInputValue.includes(",")) {
@@ -86,13 +89,12 @@ function getValue(event) {
 
   searchInput.value = "";
 
-  // Above, I log those values to the scrren and call a new function passing new city and state as parameters.
+  // Above, I log those values to the screen and call a new function passing new city and state as parameters.
 }
 
 function getLatLonData(cityUpper, stateUpper) {
-  const apiURL = `http://api.openweathermap.org/geo/1.0/direct?q=${cityUpper}&limit=5&appid=1f84100edd7f6cddf642b63a288eb2cd`;
+  const apiURL = `http://api.openweathermap.org/geo/1.0/direct?q=${cityUpper}&limit=7&appid=1f84100edd7f6cddf642b63a288eb2cd`;
   let statefound = false;
-  let cities = JSON.parse(localStorage.getItem("cities")) || [];
   // Above, I get the url  for the lon and lat from api given on page.
   fetch(apiURL)
     .then(function (response) {
@@ -104,6 +106,9 @@ function getLatLonData(cityUpper, stateUpper) {
             console.log(data);
           }
           // Above, I fetch the data and return it in json format.
+
+          let cities = JSON.parse(localStorage.getItem("cities")) || [];
+
           for (i = 0; i < data.length; i++) {
             if (data[i].state === stateUpper) {
               statefound = true;
@@ -111,49 +116,80 @@ function getLatLonData(cityUpper, stateUpper) {
               let lat = data[i].lat;
               let lon = data[i].lon;
 
-              let city = {
-                name: cityUpper,
-                state: stateUpper,
-                lati: lat,
-                long: lon,
-                rendered: "false",
-              };
-              // Above, I do a for loop for the data so it can match the exact output im looking for.
-              cities.push(city);
+              let existingCity = cities.find(
+                (city) => city.fullName === cityUpper + " " + stateUpper
+              );
+              // Above, I added this code if a city already exists it will not be re-made.
+              if (!existingCity) {
+                let city = {
+                  name: cityUpper,
+                  state: stateUpper,
+                  lati: lat,
+                  long: lon,
+                  fullName: cityUpper + " " + stateUpper
+                };
+                // Above, I do a for loop for the data so it can match the exact output im looking for.
+                // Also, I check if the city alread exists in the city array. 
+                cities.push(city);
 
-              let citiesSerialized = JSON.stringify(cities);
+                let citiesSerialized = JSON.stringify(cities);
 
-              localStorage.setItem("cities", citiesSerialized);
+                localStorage.setItem("cities", citiesSerialized);
+                //Above, I log the data and save the object created from the data to local storage.
 
-              let cityName = document.querySelector("#cityName");
-              cityName.textContent = city.name;
+                let cityName = document.querySelector("#cityName");
+                cityName.textContent = city.fullName;
+                // Above, we set the text content of the main area to the city searched. 
 
-              console.log(city);
-              getWeatherInfo(lat, lon);
+    
+
+                console.log(city);
+                createCityBtn(city);
+                getWeatherInfo(lat, lon, cityUpper, stateUpper);
+
+              // Above, I create a button for that city and call get weather info.
+                break;
+              } else {
+                console.log("City already exists:", cityUpper);
+              }
+              // Above, is a error if a city already exists.
+
             }
-            // Above, I log the data and save the object created from the data to local storage.
-            if (!statefound) {
-              alert("State Not found!");
-              break;
-            }
+          }
+
+          if (!statefound) {
+            alert("State Not found!");
           }
         });
       }
     })
-    // Above, is my alert if a state is not found
+              // Above, is my alert if a state is not found
     .catch(function (error) {
       alert("Unable to connect to server!");
     });
 
+
+
+ 
+
   // Above, is my alert if the api is not working.
+
+
 }
 
-function getWeatherInfo(lat, lon) {
+function getWeatherInfo(lat, lon, cityUpper, stateUpper) {
   const cityURL = `https://api.openweathermap.org/data/3.0/onecall?lat=${lat}&lon=${lon}&units=imperial&exclude=hourly,minutely&appid=1f84100edd7f6cddf642b63a288eb2cd`;
   // Above, I get the url to find precise location using lon and lat
+  // Also, we pass in the city and state upper parameters to to store to local storage for our last search array. 
 
   let cities = cityList;
   let cityName = document.querySelector("#cityName");
+
+  let lastSearches = JSON.parse(localStorage.getItem("lastSearches")) || [];
+  lastSearches.push({ lat: lat, lon: lon, fullName: cityUpper + " " + stateUpper});
+  localStorage.setItem("lastSearches", JSON.stringify(lastSearches));
+
+  // It is also very important I set the last state searched even if we already checked if it is in the cities array.
 
   fetch(cityURL)
     .then(function (response) {
@@ -170,6 +206,10 @@ function getWeatherInfo(lat, lon) {
     });
   // Above, is for if the api fails.
 
+
+
+
+
   /*for (i = 0; i <= cities.length; i++) {
 
     console.log(cities[i]);
@@ -182,16 +222,38 @@ function getWeatherInfo(lat, lon) {
     }
   }*/
 
-  cities.forEach((city) => {
-    if (city.lati === lat && city.long === lon) {
-      cityName.textContent = city.name;
-    }
-  });
+  //  cities.forEach((city) => {
+  //   if (city.lati === lat && city.long === lon) {
+  //     cityName.textContent = city.fullName;
+  //   }
+  // });
+
+  
+
+ 
+
+
+
+
+
+
 }
 // the commented out code above is for my refrence, it is for a bug I encountered due to looping incorrectly so values were not passed correctly.
 // solve1 = remove "=" sign in condition
 // solve2 = add conditional chaining with ?.
 // solve3 = for each loop
+
+
+
+
+
+
+
+
+
+
+
+
 
 function displayWeatherData(data) {
   console.log(data.daily);
@@ -392,27 +454,141 @@ function displayWeatherData(data) {
     // Above are  if statments to set the content of each card Humidity text content.
   });
 
-  console.log("hello");
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+  
+}
+
+function createCityBtn(city) {
+  cityBTN = document.createElement("button");
+  cityBTN.classList.add(
+    "col-12",
+    "col-sm-11",
+    "col-md-12",
+    "citiesBtn",
+    "btn-primary"
+  );
+  //Above, I created a button and gave it pre made classes,
+  cityBTN.setAttribute("type", "button");
+  cityBTN.textContent = city.name + " " + city.state;
+  cityBTN.setAttribute("data-lat", city.lati);
+  cityBTN.setAttribute("data-lon", city.long);
+
+  // Above, I set the text content and gave the button some attributes.
+
+  cityBTN.addEventListener("click", function (event) {
+    displayOldCity(event); // Pass the event object to the displayOldCity function
+  });
+
+  //Above, I added a event listener to the button  and called displayoldcity.
+
+  cityInput.append(cityBTN);
+
+  // Last, I append the button to the cityInput section.
 }
 
 function renderlastSearch() {
-  let cities = cityList;
 
-  console.log(cities);
+  let lastSearch = JSON.parse(localStorage.getItem("lastSearches")) || [];
+  let cities = cityList
+
+  
+  // Above, is where we use the last searches array. 
+ // I retrieve the cities aray to render the buttons and the last searches array to render the last search. 
+  
+
+  console.log(lastSearch);
 
   if (cities.length > 0) {
-    const lastItem = cities[cities.length - 1];
-    console.log(lastItem.long);
-    getWeatherInfo(lastItem.lati, lastItem.long);
+    cities.forEach((city) => {
+
+     createCityBtn(city)
+    });
   } else {
     console.log("No cities in the list");
+
+
   }
+
+  // Above, I create all the buttons after the page is refreshed. 
+
+
+  if (lastSearch.length > 0){
+  let lastSearchedCity = lastSearch[lastSearch.length - 1];
+
+  let cityName = document.querySelector("#cityName");
+
+  cityName.textContent = lastSearchedCity.fullName
+
+   cityNameArray = lastSearchedCity.fullName.split(" ");
+    let cityUpper = cityNameArray[0];
+   let stateUpper = cityNameArray[1];
+
+// Above,  we use the last searched city array to render the last searched city.
+
+    getWeatherInfo(lastSearchedCity.lat, lastSearchedCity.lon, cityUpper, stateUpper);
+  };
+
+// Above, we call get weather data to present the data for the last searched city. 
+  
+}
+
+function displayOldCity(event) {
+  event.preventDefault();
+  event.stopPropagation();
+
+
+ 
+
+  
+  const lat = event.target.getAttribute("data-lat");
+  const lon = event.target.getAttribute("data-lon");
+// Above, i get the lat and lon set to that buttons data attributes. 
+  
+  let cityName = document.querySelector("#cityName");
+
+  cityName.textContent = event.target.textContent;
+
+  let cityNameArray = event.target.textContent.split(" ");
+  let cityUpper = cityNameArray[0];
+  let stateUpper = cityNameArray[1];
+
+
+
+ // Above, I set the text content of the main area to the event button city name.
+ // Also, we get the sity and state upper values again.
+
+
+  getWeatherInfo(lat, lon, cityUpper,stateUpper);
+
+  // Above, I call the function get weather info to display the new data for the searched button clicked.
+  
 }
 
 $(document).ready(function () {
   // Handler for .ready() called.
+
   renderlastSearch();
+
   // I called render last search first.
   displaydate();
   searchButton.addEventListener("click", getValue);
+  // Above, I call display date to display the dates and add an event listener to the submit button. 
 });
